Fixed motor going in same direction
Turns out all I needed to do was used the motor's encoder positions and have a positive or negative position for the motor to go in different direction.

Haven't face the loopOverrun error yet but that was what made it go over its limit. It may have reset the angle Falcon integrated encoder even though its an absolute encoder.